Overhaul on redisign and Added a location API
Changed css of most of the site:
will need to work on the nav bar, footer and some more css.

Location API:
Added an api that will take users location: made it so it asks for permission before taking and then results closest city that will have the best connection.